Move things around to be more organized
