+++ nwc_backend/nostr/nostr_config.py
